Print buffer: move necessary functions to generator and remove others
